+++ python/cudf/cudf/tests/test_csv.py
@@ -522,7 +522,7 @@ def test_csv_reader_NaN_values():
     default_na_cells = (
         "#N/A\n#N/A N/A\n#NA\n-1.#IND\n"
         "-1.#QNAN\n-NaN\n-nan\n1.#IND\n"
-        "1.#QNAN\nN/A\nNA\nNULL\n"
+        "1.#QNAN\nN/A\n<NA>\nNA\nNULL\n"
         "NaN\nn/a\nnan\nnull\n"
     )
     custom_na_cells = "NV_NAN\nNotANumber\n"
@@ -541,6 +541,15 @@ def test_csv_reader_NaN_values():
         names=names,
         dtype=dtypes,
         na_values=custom_na_values,
+    )
+    assert all(np.isnan(all_nan.to_pandas()["float32"]))
+
+    # custom NA values
+    all_nan = read_csv(
+        StringIO(empty_cells + default_na_cells + "_NAA_\n"),
+        names=names,
+        dtype=dtypes,
+        na_values="_NAA_",
     )
     assert all(np.isnan(all_nan.to_pandas()["float32"]))
 
@@ -1682,3 +1691,52 @@ def test_csv_write_no_caller_manipulation():
     df_copy = df.copy(deep=True)
     _ = df.to_csv(index=True)
     assert_eq(df, df_copy)
+
+
+@pytest.mark.parametrize(
+    "df",
+    [
+        cudf.DataFrame({"a": [1, 2, 3], "": [10, 20, 40]}),
+        cudf.DataFrame({"": [10, 20, 40], "a": [1, 2, 3]}),
+        cudf.DataFrame(
+            {"a": [1, 2, 3], "": [10, 20, 40]},
+            index=cudf.Index(["a", "z", "v"], name="custom name"),
+        ),
+    ],
+)
+@pytest.mark.parametrize("index", [True, False])
+@pytest.mark.parametrize("columns", [["a"], [""], None])
+def test_csv_write_empty_column_name(df, index, columns):
+    pdf = df.to_pandas()
+    expected = pdf.to_csv(index=index, columns=columns)
+    actual = df.to_csv(index=index, columns=columns)
+
+    assert expected == actual
+
+
+@pytest.mark.parametrize(
+    "df",
+    [
+        cudf.DataFrame(),
+        cudf.DataFrame(index=cudf.Index([], name="index name")),
+    ],
+)
+@pytest.mark.parametrize(
+    "index",
+    [
+        True,
+        pytest.param(
+            False,
+            marks=pytest.mark.xfail(
+                reason="https://github.com/rapidsai/cudf/issues/6691"
+            ),
+        ),
+    ],
+)
+def test_csv_write_empty_dataframe(df, index):
+    pdf = df.to_pandas()
+
+    expected = pdf.to_csv(index=index)
+    actual = df.to_csv(index=index)
+
+    assert expected == actual